Backend/requirements.txt
langchain
langgraph
langchain-groq
langchain-community
pinecone-client
tavily-python
python-dotenv
pytest
langchain-pinecone 
langchain-huggingface 
sentence-transformers
bs4
langchain-tavily
fastapi